+++ lib/src/path/json_path.dart
@@ -1,7 +1,7 @@
-import 'package:json_path/src/grammar.dart' as grammar;
-import 'package:json_path/src/json_path_match.dart';
-import 'package:json_path/src/match_factory.dart';
-import 'package:json_path/src/selector/selector.dart';
+import 'package:json_path/src/path/grammar.dart' as grammar;
+import 'package:json_path/src/path/json_path_match.dart';
+import 'package:json_path/src/path/match_factory.dart';
+import 'package:json_path/src/path/selector/selector.dart';
 
 /// A JSONPath expression
 class JsonPath {