+++ lib/src/selector/selector.dart
@@ -1,6 +1,0 @@
-import 'package:json_path/src/json_path_match.dart';
-
-abstract class Selector {
-  /// Applies this filter to the [matches]
-  Iterable<JsonPathMatch> read(JsonPathMatch match);
-}